+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/generics/Webhook.java
@@ -5,8 +5,7 @@ import org.telegram.telegrambots.exceptions.TelegramApiRequestException;
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief TODO
- * @date 29 of October of 2016
+ * Webhook interface
  */
 public interface Webhook {
     void startServer() throws TelegramApiRequestException;