+++ telegrambots-meta/src/test/java/org/telegram/telegrambots/test/TestTelegramApi.java
@@ -9,8 +9,6 @@ import org.telegram.telegrambots.exceptions.TelegramApiRequestException;
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief TODO
- * @date 29 of October of 2016
  */
 public class TestTelegramApi extends TestBase {