+++ telegrambots-meta/src/test/java/org/telegram/telegrambots/test/asserts/ApiAssert.java
@@ -3,8 +3,6 @@ package org.telegram.telegrambots.test.asserts;
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief TODO
- * @date 04 of November of 2016
  */
 public final class ApiAssert {
     private ApiAssert() {}