+++ telegrambots-meta/src/test/java/org/telegram/telegrambots/test/base/TestBase.java
@@ -10,8 +10,6 @@ import org.telegram.telegrambots.generics.Webhook;
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief TODO
- * @date 29 of October of 2016
  */
 public abstract class TestBase {