+++ telegrambots-meta/src/test/java/org/telegram/telegrambots/test/fakes/FakeBotSession.java
@@ -7,8 +7,6 @@ import org.telegram.telegrambots.generics.LongPollingBot;
 /**
  * @author Ruben Bermudez
  * @version 1.0
- * @brief TODO
- * @date 29 of October of 2016
  */
 public class FakeBotSession implements BotSession {
     @Override